Convert to using customizer, add new type of rack
Setup customizer variables, and added a new rack in a '[' shape (note,
have not printed yet, currently untested). Added stls for 0.2ml, 2ml,
and 5ml centrifuge tubes with the new design (size may not be correct
for all tubes). New style of rack is designed to be printed without
support, upside down.